requirements.txt
Flask==3.0.3
python_docx==1.1.2
Werkzeug==3.0.3
